Update game_spec to include test coverage for guess logic
Tests indicate whether or not guesses are counted towards the correct totals

When unsaved, the guess should not be included in any logic that is used to calculate game state.
This prevents errors where the final needed guess to meet an end condition is denied by the validations
used to prevent additional guesses after the game is complete.