Offload random game creation to model
Random game creation should be processed under the standard post games route so
an additional paramater is used to determine what type of game is being created

Based off of this, either a custom game, or a random game will be created.

As the random game creation logic is specific to game, the logic has been moved
into the model.